requirements.txt
streamlit
pandas
numpy
plotly
Pillow
opencv-python
scikit-learn

# Install required packages
pip install streamlit opencv-python numpy

# Run the application
streamlit run image_caption_generation.py

